finished merging! TODO: supports showing empty value for comobobx
